+++ packages/ui/modules/index.d.ts
@@ -14,7 +14,8 @@ import {
   ConjsProps, FieldProps,
   WidgetProps, TextWidgetProps, DateTimeWidgetProps, BooleanWidgetProps, NumberWidgetProps, SelectWidgetProps, 
   TreeSelectWidgetProps, RangeSliderWidgetProps, CaseValueWidgetProps, ConfigContext, FactoryWithContext,
-  RenderedReactElement, SerializedFunction
+  RenderedReactElement, SerializedFunction,
+  Utils as CoreUtils
 } from "@react-awesome-query-builder/core";
 
 // re-export
@@ -212,6 +213,17 @@ interface VanillaWidgets {
   FuncWidget: ElementType<WidgetProps>,
 }
 
+/////////////////
+// extend Utils
+/////////////////
+
+export interface Utils extends CoreUtils {
+  // ReactUtils: {
+  //   useOnPropsChanged(obj: ReactElement): void;
+  // },
+}
+
+export declare const Utils: Utils;
 
 /////////////////