+++ packages/ui/modules/utils/reactUtils.js
@@ -66,20 +66,37 @@ export const pureShouldComponentUpdate = (self) => function(nextProps, nextState
   );
 };
 
-const canUseUnsafe = () => {
+const canUseOldComponentWillReceiveProps = () => {
   const v = React.version.split(".").map(parseInt.bind(null, 10));
-  return v[0] == 16 && v[1] >= 3 || v[0] > 16;
+  return v[0] == 16 && v[1] < 3 || v[0] < 16;
 };
 
 export const useOnPropsChanged = (obj) => {
-  if (canUseUnsafe) {
-    obj.UNSAFE_componentWillReceiveProps = (nextProps) => {
-      obj.onPropsChanged(nextProps);
-    };
-  } else {
+  // 1. `shouldComponentUpdate` should be called after `componentWillReceiveProps`
+  // 2. `shouldComponentUpdate` should not be used for PureComponent
+
+  // Because `useOnPropsChanged` can only be applied to `Component` not `PureComponent`, make it pure now
+  if (!obj.shouldComponentUpdate) {
+    obj.shouldComponentUpdate = pureShouldComponentUpdate(obj);
+  }
+
+  if (canUseOldComponentWillReceiveProps()) {
+    // Use old method
     obj.componentWillReceiveProps = (nextProps) => {
       obj.onPropsChanged(nextProps);
     };
+  } else {
+    // Simulate `componentWillReceiveProps` with `shouldComponentUpdate`
+    const origShouldComponentUpdate = obj.shouldComponentUpdate;
+    const newShouldComponentUpdate = function(nextProps, nextState) {
+      const shouldNotify = !shallowEqual(obj.props, nextProps);
+      if (shouldNotify) {
+        obj.onPropsChanged(nextProps);
+      }
+      const shouldUpdate = origShouldComponentUpdate.call(obj, nextProps, nextState);
+      return shouldUpdate;
+    };
+    obj.shouldComponentUpdate = newShouldComponentUpdate.bind(obj);
   }
 };