+++ packages/core/modules/utils/validation.js
@@ -305,17 +305,18 @@ const validateNormalValue = (leftField, field, value, valueSrc, valueType, async
     const wType = wConfig.type;
     const jsType = wConfig.jsType;
     const fieldSettings = fieldConfig.fieldSettings;
+    const listValues = fieldSettings.treeValues || fieldSettings.listValues;
 
     if (valueType && valueType != wType)
       return [`Value should have type ${wType}, but got value of type ${valueType}`, value];
-    if (jsType && !isTypeOf(value, jsType) && !fieldSettings.listValues) { //tip: can skip tye check for listValues
+    if (jsType && !isTypeOf(value, jsType) && !listValues) { //tip: can skip type check for listValues
       return [`Value should have JS type ${jsType}, but got value of type ${typeof value}`, value];
     }
 
     if (fieldSettings) {
-      const listValues = asyncListValues || fieldSettings.listValues;
-      if (listValues && !fieldSettings.allowCustomValues) {
-        return validateValueInList(value, listValues, canFix, isEndValue, config.settings.removeInvalidMultiSelectValuesOnLoad);
+      const realListValues = asyncListValues || listValues;
+      if (realListValues && !fieldSettings.allowCustomValues) {
+        return validateValueInList(value, realListValues, canFix, isEndValue, config.settings.removeInvalidMultiSelectValuesOnLoad);
       }
       if (fieldSettings.min != null && value < fieldSettings.min) {
         return [`Value ${value} < min ${fieldSettings.min}`, canFix ? fieldSettings.min : value];