+++ packages/examples/demo/config.tsx
@@ -526,7 +526,7 @@ export default (skin: string) => {
       fieldSettings: {
         treeExpandAll: true,
         // * deep format (will be auto converted to flat format):
-        // listValues: [
+        // treeValues: [
         //     { value: "1", title: "Warm colors", children: [
         //         { value: "2", title: "Red" },
         //         { value: "3", title: "Orange" }
@@ -541,7 +541,7 @@ export default (skin: string) => {
         //     ] }
         // ],
         // * flat format:
-        listValues: [
+        treeValues: [
           { value: "1", title: "Warm colors" },
           { value: "2", title: "Red", parent: "1" },
           { value: "3", title: "Orange", parent: "1" },
@@ -558,7 +558,7 @@ export default (skin: string) => {
       type: "treemultiselect",
       fieldSettings: {
         treeExpandAll: true,
-        listValues: [
+        treeValues: [
           {
             value: "1", title: "Warm colors", children: [
               { value: "2", title: "Red" },